Create integration test category, move integration tests to own namespace
