+++ CollAction.Tests/Integration/ImageServicesTests.cs
@@ -14,9 +14,10 @@ using System.Net;
 using System.Net.Http;
 using System.Threading.Tasks;
 
-namespace CollAction.Tests
+namespace CollAction.Tests.Integration
 {
     [TestClass]
+    [TestCategory("Integration")]
     public sealed class ImageServicesTests
     {
         private readonly byte[] _image = new byte[] { 0x42, 0x4D, 0x1E, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x1A, 0x00, 0x00, 0x00, 0x0C, 0x00, 0x00, 0x00, 0x01, 0x00, 0x01, 0x00, 0x01, 0x00, 0x18, 0x00, 0x00, 0x00, 0xFF, 0x00 };
